+++ py3DSkit/__init__.py
@@ -1,3 +1,4 @@
 from py3DSkit.audio import *
-from py3DSkit.graphics import *
+from py3DSkit.graphx import *
+from py3DSkit.graphp import *
 from py3DSkit.compression import *